pyproject.toml
[tool.poetry]
name = "deskassign"
version = "0.1.0"
description = "Compute a COVID-19-safe desk use schedule for Cornell CS"
authors = ["Wil Thomason <wil.thomason@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
z3 = "^0.2.0"
fire = "^0.3.1"
pyyaml = "^5.3.1"
PyPDF2 = "^1.26.0"
coloredlogs = "^14.0"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
